requirements.txt
asgiref==3.3.1
backcall==0.2.0
blessings==1.7
bpython==0.21
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
curtsies==0.3.5
cwcwidth==0.1.4
decorator==5.0.6
Django==3.1.7
django-crispy-forms==1.11.2
django-extensions==3.1.2
django-filter==2.4.0
django-phonenumber-field==5.0.0
graphviz==0.16
greenlet==1.0.0
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
livereload==2.6.3
parso==0.8.2
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
pydot==1.4.2
Pygments==2.8.1
pyparsing==2.4.7
pytz==2021.1
pyxdg==0.27
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
